refactor: update service list, remove duplicate items
