--- src/pages/ServicesPage/constants.tsx
@@ -1,59 +1,0 @@
-const ServiceItems = [
-  {
-    icon: <div className="text-3xl text-center">👓</div>,
-    href: "/market-watch",
-    title: "Market Watch",
-    description: "Overview of stock prices, indices, and market trends.",
-  },
-  {
-    icon: <div className="text-xl text-center">✔️</div>,
-    href: "/ipo-checker",
-    title: "IPO Checker",
-    description:
-      "verify whether or not you were allotted the share and the number of shares being allotted.",
-  },
-  {
-    icon: <div className="text-xl text-center">📰</div>,
-    href: "/ipo-fpo-updates",
-    title: "IPO/FPO Updates",
-    description:
-      "Timely updates on IPO and FPO to aid informed investment decisions.",
-  },
-  {
-    icon: <div className="text-xl text-center">📊</div>,
-    href: "/trading-signals",
-    title: "Trading Signals",
-    description:
-      "Receive actionable trading signals to assist in making strategic buy or sell decisions.",
-  },
-  {
-    icon: <div className="text-xl text-center">🔎</div>,
-    href: "/portfolio-tracker",
-    title: "Portfolio Tracker",
-    description:
-      "Effortlessly monitor and manage your investments with portfolio tracker.",
-  },
-  {
-    icon: <div className="text-xl text-center">📈</div>,
-    href: "/chart",
-    title: "Chart",
-    description:
-      "Visualize stock performance through dynamic and interactive charts.",
-  },
-  {
-    icon: <div className="text-xl text-center">⚖</div>,
-    href: "/compare-company",
-    title: "Compare Company",
-    description:
-      "Easily compare key financial metrics between different companies.",
-  },
-  {
-    icon: <div className="text-4xl text-center">🖩</div>,
-    href: "/share-calculator",
-    title: "Share Calculator",
-    description:
-      "Get to know the actual figures when buying and selling shares.",
-  },
-];
-
-export default ServiceItems;
